Refactor project structure: move source to src, update imports, and add initial API setup
