--- source/__init__.py
@@ -1,2 +1,0 @@
-# industriindex-de package
-__version__ = "0.1.0"
