+++ tests/test_diagnostics_controller.py
@@ -1,7 +1,7 @@
 import pytest
 from fastapi.testclient import TestClient
-from source.__main__ import app
-from source.controllers.diagnostics import healthcheck_endpoint
+from src.__main__ import app
+from src.controllers.diagnostics import healthcheck_endpoint
 
 
 client = TestClient(app)
